+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ---
 ## master
+
+## 2.13.0
 * Refactoring publication and configuration logic
 * Bumping dependencies and migration to Kotlin 1.9.10
 * Cleaning up in favor of new Kotlin/Native memory model